move schemas to separate file instead of route
