+++ backend/app/schemas/azure_blob.py
@@ -1,0 +1,26 @@
+from typing import Optional
+from pydantic import BaseModel
+
+
+class AzureConnection(BaseModel):
+    connectionstring: Optional[str] = None
+    container: Optional[str] = None
+
+
+class AzureListRequest(AzureConnection):
+    prefix: Optional[str] = None
+
+
+class AzureFileRequest(AzureConnection):
+    filename: str
+    prefix: Optional[str] = None
+    overwrite: Optional[bool] = True
+    content: Optional[str] = None
+    binary: Optional[bool] = False
+
+
+class AzureMoveRequest(AzureConnection):
+    source_name: str
+    destination_name: str
+    source_prefix: Optional[str] = None
+    destination_prefix: Optional[str] = None
